+++ src/mdns_beacon/cli/layouts.py
@@ -1,6 +1,6 @@
 """Console layout for mdns-beacon."""
 from abc import ABC, abstractmethod
-from typing import Any, Dict, Optional
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 from rich.console import RenderableType
 from rich.live import Live
@@ -17,10 +17,27 @@ class BaseLayout(ABC):
         Derived layouts must override the `renderable` property.
     """
 
+    _spinner: Optional[Spinner] = None
+
     def __init__(self, live: Live) -> None:
         """Init layout."""
         self.live = live
         self.live.update(self.renderable)
+
+    @property
+    @abstractmethod
+    def spinner_text(self) -> str:
+        """Get the spinner text to render.
+
+        Property that derived layouts must override.
+        """
+
+    @property
+    def spinner(self) -> Spinner:
+        """Spinner status annimation."""
+        if not self._spinner:
+            self._spinner = Spinner("dots", text=Text(self.spinner_text, style="green"))
+        return self._spinner
 
     @property
     @abstractmethod
@@ -34,19 +51,10 @@ class BaseLayout(ABC):
 class BlinkLayout(BaseLayout):
     """Blink cli layout."""
 
-    _spinner: Optional[Spinner] = None
+    spinner_text = "Announcing services (Press CTRL+C to quit) ..."
 
     @property
-    def spinner(self) -> Spinner:
-        """Blink spinner status annimation."""
-        if not self._spinner:
-            self._spinner = Spinner(
-                "dots", text=Text("Announcing services (Press CTRL+C to quit) ...", style="green")
-            )
-        return self._spinner
-
-    @property
-    def renderable(self) -> RenderableType:
+    def renderable(self) -> Table:
         """Blink renderable layout (spinner with status)."""
         layout = Table.grid(padding=1, expand=True)
         layout.add_row(self.spinner)
@@ -57,25 +65,43 @@ class ListenLayout(BaseLayout):
     """Listen cli layout."""
 
     services: Dict[str, Any] = {}
-    TABLE_SERVICES_COLUMNS = [
-        "#",
-        "Type",
-        "Name",
-        "Address IPv4",
-        "Port",
-        "Server",
-        "TTL",
-    ]
-    _spinner: Optional[Spinner] = None
+    TABLE_SERVICES_COLUMNS = {
+        "type": "Type",
+        "name": "Name",
+        "ipv4_address": "Address IPv4",
+        "ipv6_address": "Address IPv6",
+        "port": "Port",
+        "server": "Server",
+        "ttl": "TTL",
+        "weight": "Weight",
+        "priority": "Priority",
+        "text": "TXT",
+        "properties": "Properties",
+    }
+    DEFAULT_SHOW_COLUMNS = (
+        "type",
+        "name",
+        "ipv4_address",
+        "port",
+        "server",
+        "ttl",
+    )
 
-    @property
-    def spinner(self) -> Spinner:
-        """Listen spinner status annimation."""
-        if not self._spinner:
-            self._spinner = Spinner(
-                "dots", text=Text("Listen for services (Press CTRL+C to quit) ...", style="green")
+    spinner_text = "Listen for services (Press CTRL+C to quit) ..."
+
+    def __init__(
+        self,
+        show_columns: Optional[Union[Tuple[str], List[str]]] = None,
+        *args: Any,
+        **kwargs: Any,
+    ) -> None:
+        """Init listen layout."""
+        self.show_columns = show_columns or self.DEFAULT_SHOW_COLUMNS
+        if not set(self.show_columns).issubset(self.TABLE_SERVICES_COLUMNS.keys()):
+            raise ValueError(
+                "Unknown fields %s", set(self.show_columns) - self.TABLE_SERVICES_COLUMNS.keys()
             )
-        return self._spinner
+        super().__init__(*args, **kwargs)
 
     @property
     def services_table(self) -> Table:
@@ -88,14 +114,12 @@ class ListenLayout(BaseLayout):
             ":satellite_antenna::police_car_light:"
         )
 
-        for key in self.TABLE_SERVICES_COLUMNS:
-            table.add_column(key, no_wrap=True)
+        table.add_column("#", no_wrap=True)
+        for c in self.show_columns:
+            table.add_column(self.TABLE_SERVICES_COLUMNS[c], no_wrap=True)
 
         for index, service in enumerate(self.services.values()):
-            table.add_row(
-                str(index),
-                *[str(v) for k, v in service.items() if k in self.TABLE_SERVICES_COLUMNS],
-            )
+            table.add_row(str(index), *[str(service[c]) for c in self.show_columns])
         return table
 
     @property
@@ -117,11 +141,18 @@ class ListenLayout(BaseLayout):
             info = zeroconf.get_service_info(service_type, name)
             if info:
                 self.services[service_id] = {
-                    "Type": info.type,
-                    "Name": info.name,
-                    "Address IPv4": ",".join(info.parsed_addresses(IPVersion.V4Only)),
-                    "Port": info.port,
-                    "Server": info.server,
-                    "TTL": info.host_ttl,
+                    "type": info.type,
+                    "name": info.name,
+                    "ipv4_address": ",".join(info.parsed_addresses(IPVersion.V4Only)),
+                    "ipv6_address": ",".join(info.parsed_addresses(IPVersion.V6Only)),
+                    "port": info.port,
+                    "server": info.server,
+                    "ttl": info.host_ttl,
+                    "weight": info.weight,
+                    "priority": info.priority,
+                    "text": info.text.decode("utf8"),
+                    "properties": {
+                        k.decode("utf8"): v.decode("utf8") for k, v in info.properties.items()
+                    },
                 }
         self.live.update(self.renderable)