+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-02-13
+### Added
+- `show` option to `listen` command.
+- `weight` option to the `blink` command.
+- `priority` option to the `blink` command.
+- Mutually exclusive `txt` and `property` options to the `blink` command.
+
 ## [0.2.0] - 2021-01-22
 ### Changed
 - Blink cli layout.
@@ -16,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First release on PyPI.
 
-[Unreleased]: https://github.com/fedejaure/mdns-beacon/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/fedejaure/mdns-beacon/compare/v0.3.0...develop
+[0.3.0]: https://github.com/fedejaure/mdns-beacon/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/fedejaure/mdns-beacon/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/fedejaure/mdns-beacon/compare/releases/tag/v0.1.0